Move metadata models to arelle_interface
